requirements.txt
autopep8==1.4.3
dj-database-url==0.5.0
Django==2.1.4
django-cors-headers==2.4.0
django-filter==2.0.0
django-heroku==0.3.1
djangorestframework==3.9.0
gunicorn==19.9.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycodestyle==2.4.0
pytz==2018.7
whitenoise==4.1.2
